move block_class from state to vm
