+++ evm/vm/forks/spurious_dragon/__init__.py
@@ -1,8 +1,10 @@
 from typing import Type  # noqa: F401
+from evm.rlp.blocks import BaseBlock  # noqa: F401
 from evm.vm.state import BaseState  # noqa: F401
 
 from ..tangerine_whistle import TangerineWhistleVM
 
+from .blocks import SpuriousDragonBlock
 from .state import SpuriousDragonState
 
 
@@ -11,4 +13,5 @@ class SpuriousDragonVM(TangerineWhistleVM):
     fork = 'spurious-dragon'  # type: str
 
     # classes
+    block_class = SpuriousDragonBlock  # type: Type[BaseBlock]
     _state_class = SpuriousDragonState  # type: Type[BaseState]
